Restructure Headings & MultiSel
