+++ src/containers/SldEditor.js
@@ -3,15 +3,13 @@ import {Router, Switch, Route, Link} from "react-router-dom";
 import {useFirestore} from "react-redux-firebase";
 import {createBrowserHistory} from "history";
 import {FormattedMessage} from "react-intl";
-import Chart from "react-google-charts";
-import ReactWordcloud from "react-wordcloud";
+
 import Loading from "../components/Loading/Loading";
 import ProjNameEditor from "../components/ProjNameEditor/ProjNameEditor";
-import QRCode from "qrcode.react";
 import "./sldEditor.css";
 import ZhInput from "../components/ZhInput/ZhInput";
 import Header from "../components/Header/Header";
-
+import CurSld from "./CurSld";
 import "../lib/icons";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 
@@ -227,12 +225,7 @@ const SldEditor = props => {
           id="preview-close-btn"
           onClick={hidePreview}
         />
-        <CurSld
-          {...props}
-          sld={props.slds[props.curSldIndex]}
-          isFullscreen={isFullscreen}
-          nextSld={nextSld}
-        />
+        <CurSld {...props} isFullscreen={isFullscreen} nextSld={nextSld} />
       </div>
     </Fragment>
   );
@@ -412,7 +405,6 @@ const SldPage = props => {
           <div className="center">
             <CurSld
               {...props}
-              sld={sld}
               isFullscreen={props.isFullscreen}
               nextSld={props.nextSld}
             />
@@ -428,372 +420,6 @@ const SldPage = props => {
       </Route>
     );
   });
-};
-
-const CurSld = props => {
-  let optsArray = props.slds[props.curSldIndex].opts;
-  let resultArray = props.slds[props.curSldIndex].result;
-  let resultType = props.slds[props.curSldIndex].resType;
-  let optResult = null;
-  const colors = [
-    "#00b894",
-    "#ffeaa7",
-    "#fab1a0",
-    "#0984e3",
-    "#00cec9",
-    "#ff7675",
-    "#6c5ce7",
-    "#b2bec3",
-    "#fd79a8",
-    "#badc58",
-    "#686de0",
-    "#f0932b",
-    "#22a6b3",
-    "#f8c291",
-    "#60a3bc",
-    "#b71540",
-    "#0a3d62",
-    "#D6A2E8",
-    "#1B9CFC",
-    "#EAB543"
-  ];
-
-  let pieColors = null;
-  if (optsArray !== "") {
-    pieColors = props.slds[props.curSldIndex].opts.map((opt, index) => {
-      return {color: colors[index]};
-    });
-  }
-
-  if (optsArray !== "") {
-    optResult = props.slds[props.curSldIndex].opts.map((opt, index) => {
-      let result = resultArray[index] !== "" ? resultArray[index] : 0;
-      return [`${opt}`, result, colors[index], result];
-    });
-  } else {
-    optResult = null;
-  }
-
-  let data = [
-    [
-      "Options",
-      "Result",
-      {role: "style"},
-      {
-        sourceColumn: 0,
-        role: "annotation",
-        type: "string",
-        calc: "stringify"
-      }
-    ]
-  ].concat(optResult);
-
-  let barOptions = {
-    legend: {position: "none"},
-    height: "100%",
-    chartArea: {width: "80%", height: "75%"},
-    bar: {groupWidth: "68%"},
-    animation: {
-      duration: 1000,
-      easing: "out",
-      startup: true
-    },
-    vAxis: {
-      gridlines: {count: 0},
-      minorGridlines: {count: 0},
-      ticks: []
-    },
-    hAxis: {
-      textStyle: {
-        fontSize: 20
-      }
-    },
-    annotations: {
-      textStyle: {
-        fontSize: 20,
-        bold: true
-      }
-    }
-  };
-
-  let pieOptions = {
-    slices: pieColors,
-    animation: {
-      duration: 1000,
-      easing: "out",
-      startup: true
-    },
-    is3D: false,
-    pieHole: 0,
-    legend: {
-      position: "labeled",
-      textStyle: {
-        fontSize: 16
-      }
-    },
-    pieSliceText: "value",
-    pieSliceTextStyle: {
-      color: "#333",
-      fontSize: 16
-    }
-  };
-
-  if (props.isFullscreen === true) {
-    barOptions = {
-      legend: {position: "none"},
-      chartArea: {width: "80%", height: "70%"},
-      bar: {groupWidth: "68%"},
-      animation: {
-        duration: 1000,
-        easing: "out",
-        startup: true
-      },
-      vAxis: {
-        gridlines: {count: 0},
-        minorGridlines: {count: 0},
-        ticks: []
-      },
-      hAxis: {
-        textStyle: {
-          fontSize: 36
-        }
-      },
-      annotations: {
-        textStyle: {
-          fontSize: 36,
-          bold: true
-        }
-      }
-    };
-    pieOptions = {
-      slices: pieColors,
-      animation: {
-        duration: 1000,
-        easing: "out",
-        startup: true
-      },
-      is3D: false,
-      pieHole: 0,
-      legend: {
-        position: "labeled",
-        textStyle: {
-          fontSize: 36
-        }
-      },
-      pieSliceText: "value",
-      pieSliceTextStyle: {
-        color: "#333",
-        fontSize: 36
-      }
-    };
-  }
-
-  // To make sure the chart will be drawn only when there is an option exists
-  let chart = null;
-  if (optsArray !== "") {
-    if (resultType === "bar-chart") {
-      chart = (
-        <Chart
-          chartType="ColumnChart"
-          width="100%"
-          height="100%"
-          data={data}
-          options={barOptions}
-        />
-      );
-    } else if (resultType === "pie-chart") {
-      chart = (
-        <Chart
-          width="100%"
-          height="100%"
-          chartType="PieChart"
-          loader={<div>Your chart is loading...</div>}
-          data={data}
-          options={pieOptions}
-        />
-      );
-    }
-  }
-
-  let detailContainer = null;
-  if (props.slds[props.curSldIndex].hasQRCode) {
-    detailContainer = (
-      <div className="qr-code">
-        {/* different QRCode size depends on fullscreen or not */}
-        {props.isFullscreen === false ? (
-          <Fragment>
-            <div className="scan-to-join">
-              <FormattedMessage id="edit.scan-to-join" />
-              <FontAwesomeIcon icon={["far", "hand-point-right"]} size="lg" />
-            </div>
-            <QRCode
-              value={`https://conversa-a419b.firebaseapp.com/audi/${props.projId}`}
-              size={230}
-              imageSettings={{
-                src:
-                  "https://lh3.googleusercontent.com/ZgZLV4rnkakeFPtT14X_lz3BdDpv8kEQ6bzWvXgHw-Wj_WYqiPNSqkq2oUBBRMQiQePrmDmO6WBrRq7_bqFsQBvnTd1_vh4BVg7opHZRsvYUTRrgoL59qyzcflYMnmHy0NzjtngT4A=w400",
-                x: null,
-                y: null,
-                height: 45,
-                width: 45,
-                excavate: true
-              }}
-            />
-          </Fragment>
-        ) : (
-          <Fragment>
-            <div className="fullscreenFontSize scan-to-join">
-              <FormattedMessage id="edit.scan-to-join" />
-              <FontAwesomeIcon icon={["far", "hand-point-right"]} size="lg" />
-            </div>
-            <QRCode
-              value={`https://conversa-a419b.firebaseapp.com/audi/${props.projId}`}
-              size={500}
-              imageSettings={{
-                src:
-                  "https://lh3.googleusercontent.com/ZgZLV4rnkakeFPtT14X_lz3BdDpv8kEQ6bzWvXgHw-Wj_WYqiPNSqkq2oUBBRMQiQePrmDmO6WBrRq7_bqFsQBvnTd1_vh4BVg7opHZRsvYUTRrgoL59qyzcflYMnmHy0NzjtngT4A=w400",
-                x: null,
-                y: null,
-                height: 100,
-                width: 100,
-                excavate: true
-              }}
-            />
-          </Fragment>
-        )}
-      </div>
-    );
-  } else {
-    detailContainer = (
-      <div className="sub-heading-render">{props.slds[props.curSldIndex].subHeading}</div>
-    );
-  }
-
-  const clickFullscreen = () => {
-    if (
-      props.isFullscreen === true &&
-      props.slds !== undefined &&
-      props.curSldIndex !== undefined
-    ) {
-      props.nextSld();
-    }
-  };
-
-  let sldContent = null;
-  let openEndedResContent = null;
-
-  if (
-    props.slds[props.curSldIndex].openEndedRes !== [] &&
-    props.slds[props.curSldIndex].openEndedRes !== undefined
-  ) {
-    openEndedResContent = props.slds[props.curSldIndex].openEndedRes.map(
-      (item, index) => {
-        return (
-          <div className="open-ended-item" key={index} title={item}>
-            <p>{item}</p>
-          </div>
-        );
-      }
-    );
-  }
-
-  let tagCloudContent = null;
-  let tagResArray = Object.entries(props.slds[props.curSldIndex].tagRes);
-  let cloudData = tagResArray.map(item => {
-    return {text: item[0].toString(), value: item[1]};
-  });
-  let cloudOptions = {
-    colors: colors,
-    enableTooltip: true,
-    deterministic: true,
-    fontFamily: "Noto Sans TC",
-    fontSizes: [20, 60],
-    fontStyle: "normal",
-    fontWeight: "normal",
-    padding: 1,
-    rotations: 3,
-    rotationAngles: [0, 0],
-    scale: "log",
-    spiral: "archimedean",
-    transitionDuration: 1000
-  };
-  if (props.isFullscreen === true) {
-    cloudOptions = {
-      colors: colors,
-      enableTooltip: true,
-      deterministic: true,
-      fontFamily: "Noto Sans TC",
-      fontSizes: [30, 100],
-      fontStyle: "normal",
-      fontWeight: "normal",
-      padding: 1,
-      rotations: 3,
-      rotationAngles: [0, 0],
-      scale: "log",
-      spiral: "archimedean",
-      transitionDuration: 1000
-    };
-  }
-  tagCloudContent = <ReactWordcloud options={cloudOptions} words={cloudData} />;
-
-  if (props.slds[props.curSldIndex].sldType === "multiple-choice") {
-    sldContent = (
-      <Fragment>
-        <div className="qus-div">{props.slds[props.curSldIndex].qContent}</div>
-        {chart}
-      </Fragment>
-    );
-  } else if (props.slds[props.curSldIndex].sldType === "heading-page") {
-    sldContent = (
-      <div className="heading-render-container">
-        <div className="heading-render">{props.slds[props.curSldIndex].heading}</div>
-        <Fragment>{detailContainer}</Fragment>
-      </div>
-    );
-  } else if (props.slds[props.curSldIndex].sldType === "open-ended") {
-    sldContent = (
-      <Fragment>
-        <div className="qus-div">{props.slds[props.curSldIndex].qContent}</div>
-        <div id="open-sld-content">{openEndedResContent}</div>
-      </Fragment>
-    );
-  } else if (props.slds[props.curSldIndex].sldType === "tag-cloud") {
-    sldContent = (
-      <Fragment>
-        <div className="qus-div">{props.slds[props.curSldIndex].qContent}</div>
-        <div id="cloud-sld-content">{tagCloudContent}</div>
-      </Fragment>
-    );
-  }
-
-  let handRaised = null;
-  if (props.slds[props.curSldIndex].sldType !== "heading-page") {
-    handRaised = (
-      <div className="hand-group">
-        <FontAwesomeIcon icon={["fas", "hand-paper"]} id="hand-icon" />
-        {props.respondedAudi[props.slds[props.curSldIndex].id]
-          ? props.respondedAudi[props.slds[props.curSldIndex].id].length
-          : "0"}
-      </div>
-    );
-  } else {
-    handRaised = "";
-  }
-
-  return (
-    <div id="current-sld-container">
-      <div id="current-sld-border">
-        <div id="current-sld" onClick={clickFullscreen}>
-          {sldContent}
-          <div className="member-info">
-            {handRaised}
-            <FontAwesomeIcon icon={["fas", "user"]} id="member-icon" />
-            {props.involvedAudi.length}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
 };
 
 const AddSldBtn = props => {
