+++ src/lib/colors.js
@@ -1,0 +1,24 @@
+const colors = [
+  "#00b894",
+  "#ffeaa7",
+  "#fab1a0",
+  "#0984e3",
+  "#00cec9",
+  "#ff7675",
+  "#6c5ce7",
+  "#b2bec3",
+  "#fd79a8",
+  "#badc58",
+  "#686de0",
+  "#f0932b",
+  "#22a6b3",
+  "#f8c291",
+  "#60a3bc",
+  "#b71540",
+  "#0a3d62",
+  "#D6A2E8",
+  "#1B9CFC",
+  "#EAB543"
+];
+
+export default colors;
